ostree-kernel-initramfs: add a condition check
A regression was introduced by 3f3faa9f:
[ ostree-kernel-initramfs: add ostree-devicetree-overlays package ]

Check if ${DEPLOY_DIR_IMAGE}/overlays.txt present before installing it,
not all machines deploys overlays.txt to ${DEPLOY_DIR_IMAGE}.

Related-to: TOR-1700

+++ recipes-sota/ostree-kernel-initramfs/ostree-kernel-initramfs_%.bbappend
@@ -10,8 +10,10 @@ do_install_append () {
             install -m 0644 ${DEPLOY_DIR_IMAGE}/overlays.txt $kerneldir/dtb
         fi
     elif [ "${KERNEL_IMAGETYPE}" = "fitImage" ]; then
-        install -d $kerneldir/dtb
-        install -m 0644 ${DEPLOY_DIR_IMAGE}/overlays.txt $kerneldir/dtb
+        if [ -e ${DEPLOY_DIR_IMAGE}/overlays.txt ]; then
+            install -d $kerneldir/dtb
+            install -m 0644 ${DEPLOY_DIR_IMAGE}/overlays.txt $kerneldir/dtb
+        fi
     fi
 }
 do_install[depends] += "${@'virtual/dtb:do_deploy' if '${PREFERRED_PROVIDER_virtual/dtb}' else ''}"
